Rename t-only-literal to no-dyntmaic-translations
