+++ packages/eslint-plugin-sentry/lib/rules/no-dynamic-translations.js
